requirements.txt
numpy==1.20.3
pandas==1.3.3
matplotlib==3.4.3
scikit-learn==0.24.2
streamlit==1.4.0
biopython==1.78